Hw6 email and images

+++ src/services/contact.js
@@ -1,6 +1,6 @@
-import ContactCollection from '../db/models/Contact.js';
-import { calcPaginationData } from '../utils/calcPaginationData.js';
 import { sortList } from '../constants/index.js';
+import ContactCollection from '../models/Contact.js';
+import { calcPaginationData } from '../utils/calcPaginationData.js';
 
 export const getContacts = async ({
   page = 1,
@@ -10,23 +10,24 @@ export const getContacts = async ({
   filters = {},
 }) => {
   const skip = (page - 1) * perPage;
-  const contactQuery = ContactCollection.find();
+  const contactQuery = ContactCollection.find({ userId: filters.userId });
 
-  if (filters.type) {
-    contactQuery.where('contactType').equals(filters.type);
-  }
-
-  if (filters.isFavourite) {
+  if (typeof filters.isFavourite === 'boolean') {
     contactQuery.where('isFavourite').equals(filters.isFavourite);
   }
-  const totalItems = await ContactCollection.find()
-  .merge(contactQuery)
-  .countDocuments();
+  if (filters.contactType) {
+    contactQuery.where('contactType').equals(filters.contactType);
+  }
 
   const data = await contactQuery
     .skip(skip)
     .limit(perPage)
-    .sort({ [sortBy]: sortOrder });
+    .sort({ [sortBy]: sortOrder })
+    .merge(contactQuery);
+
+  const totalItems = await ContactCollection.find({
+    userId: filters.userId,
+  }).countDocuments();
 
   const paginationData = calcPaginationData({ page, perPage, totalItems });
 
@@ -39,21 +40,33 @@ export const getContacts = async ({
   };
 };
 
-export const getContactById = (id) => ContactCollection.findOne({ _id: id });
+export const getContactById = (contactId, userId) =>
+  ContactCollection.findOne({ _id: contactId, userId });
 
-export const addContact = (data) => ContactCollection.create(data);
+export const addContact = (payload) => ContactCollection.create(payload);
 
-export const updateContact = async (_id, data, options = {}) => {
+export const updateContact = async (
+  { contactId, userId },
+  payload,
+  options = {},
+) => {
   const { upsert = false } = options;
-  const rawResult = await ContactCollection.findOneAndUpdate({ _id }, data, {
-    upsert,
-    includeResultMetadata: true,
-  });
+  const rawResult = await ContactCollection.findOneAndUpdate(
+    { _id: contactId, userId },
+    payload,
+    {
+      upsert,
+      includeResultMetadata: true,
+    },
+  );
 
   if (!rawResult || !rawResult.value) return null;
 
-  return rawResult.value;
+  return {
+    data: rawResult.value,
+    isNew: Boolean(rawResult.lastErrorObject.upserted),
+  };
 };
 
-export const deleteContactById = (_id) =>
-  ContactCollection.findOneAndDelete({ _id });
+export const deleteContactById = (contactId, userId) =>
+  ContactCollection.findOneAndDelete({ _id: contactId, userId });
